requirements.txt
apache-airflow==2.9.0 
pandas==2.3.3 
numpy==2.2.6 
requests==2.32.3
python-dotenv==1.0.1
